requirements.txt
fuzzywuzzy==0.18.0
GeoAlchemy2==0.14.1
loguru==0.5.3
pandas==1.4.1
PyYAML==6.0.1
SQLAlchemy==1.4.32
SQLAlchemy_Utils==0.41.1
streamlit==1.24.1
